+++ client/src/components/DatePicker/DatePicker.css
@@ -1,0 +1,3 @@
+.boxx {
+    display: flex;
+}
