+++ client/src/components/TimeOffTable/TimeOffTable.jsx
@@ -1,0 +1,19 @@
+import React from "react";
+
+
+const TimeOffTable = () => {
+  
+  
+    return (
+      <div>
+      
+      Aquí ponemos la tabla donde se van a ver las ausencias.
+      Puedes usar la funcion de getMyRequests para obtener los datos a mostrar en la tabla.
+      La función está en  service/calendarRequest para orientarte puedes mirar la pagina calender.
+        
+      </div>
+    );
+  };
+  
+  export default TimeOffTable;
+
